Separate some functionality from immutable trait
Having a separate trait for protected magic methods is useful when we
want to have a value object without array-based value manipulation.
